scripts: implement nfs-export-updater tool for bitbake
This tool is an approach to support nfsroot-based development workflows
for on-target testing with Yocto/OE.

The idea behind this is that you export your bitbake-generated root file
system over NFS to let your target boot it.
This tool then allows you to dynamically update the contents of the
exported rootfs on a per-package base without having to re-generate and
re-export the entire root file system. This should speed up iterations
for on-target testing.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>